# fix(commands): add missing index on commands table, for performances